Smart college bus transportation system for (covid 19) situation.
In this system we check the health condition of the students before allowing them to enter the bus. If any student is found to be having the symptoms, an alert information will be passed to the management. In addition to this, the students can also get the location of the bus frequently before the bus reaches their boarding point. The system also monitors the bus attendance and inform it periodically to the management.