Hover Board
It is a self balancing board which is powered by battery and allowed riders to move across the ground without pedalling with their feet. In this project i have used two pushing switches, 12Dc motor and 12v rechargeable battery. if we push two switches at the same time it will move forward. if we want turn we can operate the  switches accordingly. If we want to turn left we can push only right side switch, at this point the left side wheel remains rest subsequently the right side wheel in motion and thus the hover board turns to the left, the same procedure follows for to turn the right side by pushing left side switch. Thus we can achieve navigation through simple switches.